move mappability code out of utils
